* Creado el create_equipment_cases.py para mayor comodidad en el manejo de la interfaz. * Modificado el create_equipment_view.py (de nombre y añadida funcionalidad) * Cambiado el nombre de maintenance_form a create_maintenance_view.py * Cambiado el nombre de piece_form a create_piece_view.py * Modificado el método create_equipment en equipment_controller.py para que retorne el nuevo enum create_equipment_cases.py * Modificado el equipment_sql.py para que verifique que todos los campos poseen datos * Modificado el login_view.py para que rediriga a una página distinta dependiendo del nivel de acceso del usuario, así como añadir en el page.session el valor local_user * Optimizado el supervisor_view.py y añadida funcionalidad * Elimiando import innecesario en welcome_page_view.py

+++ src/controllers/sql/equipment_sql.py
@@ -6,6 +6,8 @@ class EquipmentSQL(BaseSqlController):
 
     def create_equipment(self, *, code:str, name:str, description:str, provider:str) -> bool:
         """Metodo para ingresar una nueva fila en la columna de equipos"""
+        if not code or not name or not description or not provider:
+            return False
         return self._execute(
             query= f"""
             INSERT INTO {self.table()} (code, name, description, provider)
@@ -14,23 +16,27 @@ class EquipmentSQL(BaseSqlController):
             params= (code, name, description, provider))
 
     def get_by_partial_name(self, name:str) -> list[dict]:
+        if not name:
+            return []
         """Metodo que retorna todas las instancias posibles en base a un nombre parcial"""
         query= f"SELECT * FROM {self.table()} WHERE LOWER(name) LIKE ?"
         params = (f"%{name.lower()}%",)
         rows = self._fetchall(query= query, params= params)
         return [dict(row) for row in rows]
 
-    def delete_by_id(self, equipment_id:str) -> bool:
+    def delete_by_code(self, equipment_code:str) -> bool:
+        if not equipment_code:
+            return False
         """Elimina de la tabla el elemento con el id ingresado"""
         return self._execute(
             query= f"DELETE FROM {self.table()} WHERE code = ?",
-            params= (equipment_id, )
+            params= (equipment_code,)
         )
 
-    def fetch_equipment_data(self, equipment_id:str) -> dict | None:
+    def fetch_equipment_data(self, equipment_code:str) -> dict | None:
         data_fetch = self._fetchone(
             query=f"SELECT * FROM {self.table()} WHERE code = ?",
-            params= (equipment_id, )
+            params= (equipment_code,)
         )
         return dict(data_fetch) if data_fetch else None
